socket: fix the return type of quic_outq_wspace
Moritz found a call trace in a test:

  [] watchdog: BUG: soft lockup - CPU#8 stuck for 26s! [http3_test:25273]
  [] RIP: 0010:quic_frame_stream_append+0x2b7/0x3b0 [quic]
  []  quic_outq_stream_append+0x6d/0x100 [quic]
  []  quic_sendmsg+0x4c6/0x8e0 [quic]
  []  __sock_sendmsg+0x58/0x70
  []  ____sys_sendmsg+0x277/0x2f0
  []  ___sys_sendmsg+0x9a/0xe0
  []  __sys_sendmsg+0x76/0xc0
  []  do_syscall_64+0x5b/0xc0

It was caused by the return value overflow of quic_outq_wspace().
This could happen if the peer set a window (initial_max_data in
transport parameter) > U32_MAX.

Fix it by change the return type from u32 to u64.

While at it, return 0 if it is not able to append the data in
quic_frame_stream_append() so that it will not change 'bytes'
in quic_sendmsg().

Reported-by: mbuhl <git@moritzbuhl.de>

+++ modules/net/quic/output.c
@@ -145,14 +145,14 @@ int quic_outq_flow_control(struct sock *sk, struct quic_stream *stream, u16 byte
 	return blocked;
 }
 
-u32 quic_outq_wspace(struct sock *sk, struct quic_stream *stream)
+u64 quic_outq_wspace(struct sock *sk, struct quic_stream *stream)
 {
 	struct quic_outqueue *outq = quic_outq(sk);
-	u32 len = outq->max_bytes - outq->bytes;
+	u64 len = outq->max_bytes - outq->bytes;
 
 	if (stream) {
-		len = min_t(u32, len, sk_stream_wspace(sk));
-		len = min_t(u32, len, stream->send.max_bytes - stream->send.bytes);
+		len = min_t(u64, len, sk_stream_wspace(sk));
+		len = min_t(u64, len, stream->send.max_bytes - stream->send.bytes);
 	}
 
 	return len;
